requirements.txt
annotated-types==0.7.0
anyio==4.3.0
bcrypt==4.1.3
boto3==1.34.111
botocore==1.34.111
certifi==2024.2.2
click==8.1.7
colorama==0.4.6
databases==0.9.0
dnspython==2.6.1
ecdsa==0.19.0
email_validator==2.1.1
exceptiongroup==1.2.1
fastapi==0.111.0
fastapi-cli==0.0.4
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
idna==3.7
iniconfig==2.0.0
itsdangerous==2.2.0
Jinja2==3.1.4
jmespath==1.0.1
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
orjson==3.10.3
packaging==24.0
passlib==1.7.4
pluggy==1.5.0
pyasn1==0.6.0
pydantic==2.7.1
pydantic_core==2.18.2
Pygments==2.18.0
pytest==8.2.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.9
PyYAML==6.0.1
rich==13.7.1
rsa==4.9
s3transfer==0.10.1
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.30
starlette==0.37.2
tomli==2.0.1
typer==0.12.3
typing_extensions==4.11.0
ujson==5.10.0
urllib3==2.2.1
uvicorn==0.29.0
watchfiles==0.21.0
websockets==12.0
aiohttp==3.9.1
aioredlock==0.7.3
async-sqlalchemy==1.0.0
aiocache==0.12.2
aiofiles==24.1.0
aiosignal==1.2.0
pydantic-settings==2.3.4
dependency-injector==4.41.0
asyncmy==0.2.9
starlette==0.37.2
starlette-context==0.3.6
requests==2.31.0
aiosmtplib==3.0.2
aioredis==1.3.1
hiredis==2.3.2
redis==5.0.3